requirements.txt
flask
flask-socketio
werkzeug
flask-sqlalchemy
psycopg2-binary
python-dotenv
